"Fix" exploding upon spawn
Sometimes there would be a bug where you'd get crushed and die as a result.

Believe this came about only in 2.2.9.

Just spawn the player like normal and spawn the rock at the same position, THEN adjust the height of the player to be on the rock, dummy.
player.mo.floorz was being weird causing issues such as https://cdn.discordapp.com/attachments/817964598685859863/840991183559852032/srb20270.png